notmuch.el: Make notmuch-show buffer name first subject, instead of thread-id (supersedes V1 and V2)
Change the buffer name to a uniquified subject of the thread (i.e. the
subject of the first message in the thread) instead of the thread-id. This
is more meaningful to the user, and will make it easier to scroll through
numerous open buffers.

Note that this patch adds an optional `buffer-name' argument to notmuch
show.

This version supersedes V1 and V2 of this patch. It is rebased on HEAD,
ensures that the buffer names are unique, and that the `notmuch-show'
command can still be used interactively (fixing a clumsy bug in V2 which
prevented that).

Signed-off-by: Jesse Rosenthal <jrosenthal@jhu.edu>